+++ haskell.nix
@@ -13,41 +13,30 @@ in self-hs: super-hs:
       pkgs.haskell.lib.dontCheck
         (self-hs.callCabal2nix x y {});
 
+    # FIXME: hevm is broken in our current nixpkgs pin, so we pull a newer one
+    # here and then use that to build hevm-0.50.0. This should be removed once
+    # we migrate the main nixpkgs pin to a newer version.
+    pkgs-2022-11 = import (pkgs.fetchFromGitHub {
+      owner = "nixos";
+      repo = "nixpkgs";
+      rev = "dac57a4eccf1442e8bf4030df6fcbb55883cb682";
+      sha256 = "sha256-C15oAtyupmLB3coZY7qzEHXjhtUx/+77olVdqVMruAg=";
+    }) {};
+    myHaskell = pkgs-2022-11.haskellPackages.override {
+      overrides = self: super: {
+        hevm = pkgs.haskell.lib.dontCheck (self.callHackageDirect {
+          pkg = "hevm";
+          ver = "0.50.0";
+          sha256 = "sha256-ju/ZuacGneQR6tJLv7gwyMj7+u8GGQ5JcYm/XXi53yI=";
+        } {});
+        secp256k1 = pkgs.secp256k1;
+      };
+    };
+
   in {
     restless-git = dontCheck "restless-git" (./src/restless-git);
 
-    hevm = pkgs.haskell.lib.dontHaddock ((
-      self-hs.callCabal2nix "hevm" (./src/hevm) {
-        # Haskell libs with the same names as C libs...
-        # Depend on the C libs, not the Haskell libs.
-        # These are system deps, not Cabal deps.
-        inherit (pkgs) secp256k1;
-      }
-    ).overrideAttrs (attrs: {
-      postInstall =
-        if wrapped
-        then
-          ''
-            wrapProgram $out/bin/hevm --prefix PATH \
-              : "${lib.makeBinPath (with pkgs; [bash coreutils git solc])}"
-          ''
-        else "";
-
-      enableSeparateDataOutput = true;
-      buildInputs = attrs.buildInputs ++ [pkgs.solc] ++ (if wrapped then [] else [pkgs.z3 pkgs.cvc4]);
-      nativeBuildInputs = attrs.nativeBuildInputs ++ [pkgs.makeWrapper];
-      configureFlags = attrs.configureFlags ++ [
-          "--ghc-option=-O2"
-      ] ++
-      (if stdenv.isDarwin then [] else
-        if shared then [] else [
-          "--enable-executable-static"
-          "--extra-lib-dirs=${pkgs.gmp.override { withStatic = true; }}/lib"
-          "--extra-lib-dirs=${pkgs.glibc.static}/lib"
-          "--extra-lib-dirs=${pkgs.libff.override { enableStatic = true; }}/lib"
-          "--extra-lib-dirs=${pkgs.ncurses.override {enableStatic = true; }}/lib"
-          "--extra-lib-dirs=${pkgs.zlib.static}/lib"
-          "--extra-lib-dirs=${pkgs.libffi.overrideAttrs (old: { dontDisableStatic = true; })}/lib"
-      ]);
-    }));
+    eth-utils = pkgs.haskell.lib.dontHaddock (
+      myHaskell.callCabal2nix "eth-utils" (./src/eth-utils) {}
+    );
   }
